Fixed more linting errors
