+++ tests/conftest.py
@@ -1,19 +1,16 @@
 """configuration file for deltacov tests"""
 
-import pytest
-import argparse
-import pkg_resources
-import subprocess
-import xml.dom
 from xml.dom import minidom
-from pkg_resources import DistributionNotFound, VersionConflict
+import pytest
+import pkg_resources
+from pkg_resources import DistributionNotFound
 
 __version__ = '1.0.0'
 
 pytest_plugins = ["pytester"]
 
 dependencies = [
-    'pytester' , 'pytest-cov>=2.6.0'
+    'pytester', 'pytest-cov>=2.6.0'
 ]
 
 # pytest-cov
@@ -30,15 +27,13 @@ def pytest_addoption(parser):
         '--delta',
         action='store',
         default=True
-        help='pytest-deltacov help:\n\
-         --delta : Display a graph of code coverage over time (previous ten runs)'
     )
 
 
 def run_subprocess():
     """ Runs a subprocess to create the xml file """
     cmd = "pipenv run pytest -x -s --cov-config pytest.cov --cov-report xml:cov.xml --cov"
-    run.subprocess(cmd.split(), capture_output=False, shell=False)
+    subprocess(cmd.split(), capture_output=False, shell=False)
 
 
 def parse_xml_for_coverage():
@@ -54,7 +49,19 @@ def parse_xml_for_coverage():
             print("100% coverage")
 
 
-def pytest_report_header(config):
+def pytest_report_header():
     """ Prints to the header """
-    str = "Using --delta coverage report"
-    return str
+    header = "Using --delta coverage report"
+    return header
+
+
+@pytest.fixture
+def simpletestdir(testdir):
+    """ Fixture that generates simple tests for functionality checks """
+    testdir.makepyfile(
+        test_one="""
+        def test_a():
+            assert True
+        """
+    )
+    yield testdir
